+++ src/main/resources/templates/member/tutor.html
@@ -60,47 +60,27 @@
     </section>
     <section class="search">
         <h2>튜터 찾기</h2>
-        <form action="#">
+        <form th:action="@{/member/tutor}" method="get">
             <input type="text" name="search" placeholder="검색">
             <select name="region">
                 <option value="전체">전체</option>
-                <option value="서울">서울</option>
-                <option value="경기">경기</option>
-                <option value="인천">인천</option>
-                <option value="강원">강원</option>
-                <option value="충청북도">충청북도</option>
-                <option value="충청남도">충청남도</option>
-                <option value="전라북도">전라북도</option>
-                <option value="전라남도">전라남도</option>
-                <option value="경상북도">경상북도</option>
-                <option value="경상남도">경상남도</option>
-                <option value="제주특별자치도">제주특별자치도</option>
+                <!-- 나머지 옵션들... -->
             </select>
             <select name="subject" placeholder="과목">
                 <option value="전체">전체</option>
-                <option value="국어">국어</option>
-                <option value="영어">영어</option>
-                <option value="수학">수학</option>
-                <option value="과학">과학</option>
-                <option value="사회">사회</option>
-                <option value="예술">예술</option>
-                <option value="기타">기타</option>
-            </select>
-            <select name="school" placeholder="대학교">
-                <option value="전체">전체</option>
-                <option value="삼육대학교">삼육대학교</option>
-                <option value="서울대학교">서울대학교</option>
-                <option value="연세대학교">연세대학교</option>
-                <option value="고려대학교">고려대학교</option>
-                <option value="카이스트">카이스트</option>
-                <option value="포항공대">포항공대</option>
-                <option value="한양대학교">한양대학교</option>
-                <option value="중앙대학교">중앙대학교</option>
+                <!-- 나머지 옵션들... -->
             </select>
             <select name="memberGender" placeholder="성별">
                 <option value="전체">전체</option>
                 <option value="M">남자</option>
                 <option value="F">여자</option>
+            </select>
+            <select name="tutorUni" placeholder="대학교">
+                <option value="전체">전체</option>
+                <!-- 나머지 옵션들... -->
+            </select>
+            <select name="myKey" placeholder="My Key">
+                <!-- 나머지 옵션들... -->
             </select>
             <input type="submit" value="검색하기">
         </form>
@@ -135,7 +115,6 @@
         </tr>
         </tbody>
     </table>
-    </section>
 </main>
 </body>
 </html>
